+++ DATA/days.json
@@ -1,30 +1,23 @@
 [
  {
-  "id": 1,
   "day": 1
  },
  {
-  "id": 2,
   "day": 2
  },
  {
-  "id": 3,
   "day": 3
  },
  {
-  "day": 4,
-  "id": 4
+  "day": 4
  },
  {
-  "day": 5,
-  "id": 5
+  "day": 5
  },
  {
-  "day": 6,
-  "id": 6
+  "day": 6
  },
  {
-  "day": 1,
-  "id": 7
+  "day": 1
  }
 ]
