+++ pages/CreateWord.jsx
@@ -1,27 +1,31 @@
 import React, { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import clientPromise from "../lib/mongodb";
-import { Loading } from "@nextui-org/react";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
 //
 const CreateWord = ({ days, words }) => {
+	/* Destructuring the useForm hook for form validation. */
 	const {
 		register,
 		reset,
 		formState: { errors },
 		handleSubmit,
 	} = useForm();
+
+	/* Setting the state of the component. */
 	const [wordsState, setWordsState] = useState([]);
 	const [korean, setKorean] = useState("");
 	const [loading, setLoading] = useState(false);
 	const [english, setEnglish] = useState("");
 	const [dayState, setDayState] = useState("");
+
 	useEffect(() => {
 		setWordsState(words);
 	}, [words]);
 
+	/* Toastify function - a Toast that shows up when the form is submitted succesfully. */
 	const promise = () => {
 		const resolveAfter3Sec = new Promise((resolve) => setTimeout(resolve, 1000));
 		toast.promise(resolveAfter3Sec, {
@@ -43,6 +47,7 @@ const CreateWord = ({ days, words }) => {
 	const submitForm = async (data, e) => {
 		e.target.reset();
 		setLoading(true);
+		/* Sending the data to the server. */
 		let res = await fetch("http://localhost:3005/api/words", {
 			method: "POST",
 			body: JSON.stringify({
@@ -53,6 +58,7 @@ const CreateWord = ({ days, words }) => {
 				isMemorized: false,
 			}),
 		});
+		/* Checking if the response is ok, if it is, it will reset the form, and set the state to empty. */
 		if (res.ok) {
 			promise();
 			setLoading(false);
@@ -73,7 +79,9 @@ const CreateWord = ({ days, words }) => {
 							className="form-label inline-block mb-2 text-gray-700">
 							Korean
 						</label>
+						{/* A form input that is using the react-hook-form library. */}
 						<input
+							/* A spread operator that is used to pass the props to the input element and check if the Korean word was added if not show required message.*/
 							{...register("kore", {
 								required: "Korean word required.",
 								onChange: (e) => setKorean(e.target.value),
@@ -83,6 +91,7 @@ const CreateWord = ({ days, words }) => {
 							className="form-control block w-full px-3 py-1.5 text-base font-normal text-gray-700 bg-white bg-clip-padding border border-solid border-gray-300 rounded transition ease-in-out m-0 focus:text-gray-700 focus:bg-white focus:border-blue-600 focus:outline-none"
 							placeholder="컴퓨터"
 						/>
+						{/* required message */}
 						<p>{errors.kore?.message}</p>
 					</div>
 					<div className="form-group mb-6">
@@ -92,6 +101,7 @@ const CreateWord = ({ days, words }) => {
 							English
 						</label>
 						<input
+							/* A spread operator that is used to pass the props to the input element and check if the English word was added if not show required message.*/
 							{...register("engl", {
 								required: "English word required.",
 								onChange: (e) => setEnglish(e.target.value),
@@ -114,6 +124,7 @@ const CreateWord = ({ days, words }) => {
         focus:text-gray-700 focus:bg-white focus:border-blue-600 focus:outline-none"
 							placeholder="Computer"
 						/>
+						{/* required message */}
 						<p>{errors.engl?.message}</p>
 					</div>
 					<div className="form-group mb-6 flex flex-col items-center justify-center">
@@ -122,6 +133,7 @@ const CreateWord = ({ days, words }) => {
 							onChange={(e) => setDayState(e.target.value)}
 							className="form-control block w-full px-3 py-1.5 text-base font-normal text-gray-700 bg-white bg-clip-padding  rounded border-none drop-shadow-sm  transition ease-in-out m-0 focus:text-gray-700 focus:bg-white focus:border-blue-600 focus:outline-none"
 							value={dayState}>
+							{/* Mapping through the days array and returning the day number. */}
 							{days.map((dayz, dayzIdx) => {
 								return (
 									<option
@@ -189,11 +201,17 @@ export async function getServerSideProps(context) {
 			"Content-Type": "application/json",
 		},
 	});
+	/* Getting the data from the server and setting it to the words variable. */
 	let words = await res.json();
+	/* Connecting to the MongoDB database. */
 	const client = await clientPromise;
+	/* Connecting to the dictionary database. */
 	const db = client.db("dictionary");
+	/* Getting the data from the days collection and setting it to the days variable. */
 	const days = await db.collection("days").find({}).toArray();
+	/* Converting the data from the database to a JSON object. */
 	const daysdata = JSON.parse(JSON.stringify(days));
+	/* Returning the data from the database to the component. */
 	return {
 		props: { days: daysdata, words },
 	};
